api: Update tests for the new package
