+++ pkg/api/query/v1_test.go
@@ -21,10 +21,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"reflect"
 	"strings"
@@ -32,9 +30,6 @@ import (
 	"time"
 
 	"github.com/fortytw2/leaktest"
-	"github.com/go-kit/kit/log"
-	"github.com/opentracing/opentracing-go"
-	"github.com/pkg/errors"
 	"github.com/prometheus/common/route"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/timestamp"
@@ -43,9 +38,9 @@ import (
 	"github.com/prometheus/prometheus/rules"
 	"github.com/prometheus/prometheus/storage"
 
+	baseAPI "github.com/thanos-io/thanos/pkg/api"
 	"github.com/thanos-io/thanos/pkg/compact"
 	"github.com/thanos-io/thanos/pkg/component"
-	extpromhttp "github.com/thanos-io/thanos/pkg/extprom/http"
 	"github.com/thanos-io/thanos/pkg/gate"
 	"github.com/thanos-io/thanos/pkg/query"
 	"github.com/thanos-io/thanos/pkg/rules/rulespb"
@@ -109,7 +104,10 @@ func TestEndpoints(t *testing.T) {
 
 	now := time.Now()
 	timeout := 100 * time.Second
-	api := &API{
+	api := &QueryAPI{
+		baseAPI: &baseAPI.BaseAPI{
+			Now: func() time.Time { return now },
+		},
 		queryableCreate: query.NewQueryableCreator(nil, nil, store.NewTSDBStore(nil, nil, db, component.Query, nil), 2, timeout),
 		queryEngine: promql.NewEngine(promql.EngineOpts{
 			Logger:     nil,
@@ -117,19 +115,18 @@ func TestEndpoints(t *testing.T) {
 			MaxSamples: 10000,
 			Timeout:    timeout,
 		}),
-		now:  func() time.Time { return now },
 		gate: gate.NewKeeper(nil).NewGate(4),
 	}
 
 	start := time.Unix(0, 0)
 
 	var tests = []struct {
-		endpoint ApiFunc
+		endpoint baseAPI.ApiFunc
 		params   map[string]string
 		query    url.Values
 		method   string
 		response interface{}
-		errType  ErrorType
+		errType  baseAPI.ErrorType
 	}{
 		{
 			endpoint: api.query,
@@ -398,7 +395,7 @@ func TestEndpoints(t *testing.T) {
 				"query": []string{"0.333"},
 				"dedup": []string{"sdfsf"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.queryRange,
@@ -430,7 +427,7 @@ func TestEndpoints(t *testing.T) {
 				"end":   []string{"2"},
 				"step":  []string{"1"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.queryRange,
@@ -439,7 +436,7 @@ func TestEndpoints(t *testing.T) {
 				"start": []string{"0"},
 				"step":  []string{"1"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.queryRange,
@@ -448,7 +445,7 @@ func TestEndpoints(t *testing.T) {
 				"start": []string{"0"},
 				"end":   []string{"2"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		// Bad query expression.
 		{
@@ -457,7 +454,7 @@ func TestEndpoints(t *testing.T) {
 				"query": []string{"invalid][query"},
 				"time":  []string{"1970-01-01T01:02:03+01:00"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.queryRange,
@@ -467,7 +464,7 @@ func TestEndpoints(t *testing.T) {
 				"end":   []string{"100"},
 				"step":  []string{"1"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		// Invalid step.
 		{
@@ -478,7 +475,7 @@ func TestEndpoints(t *testing.T) {
 				"end":   []string{"2"},
 				"step":  []string{"0"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		// Start after end.
 		{
@@ -489,7 +486,7 @@ func TestEndpoints(t *testing.T) {
 				"end":   []string{"1"},
 				"step":  []string{"1"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		// Start overflows int64 internally.
 		{
@@ -500,7 +497,7 @@ func TestEndpoints(t *testing.T) {
 				"end":   []string{"1489667272.372"},
 				"step":  []string{"1"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		// Bad dedup parameter.
 		{
@@ -512,7 +509,7 @@ func TestEndpoints(t *testing.T) {
 				"step":  []string{"1"},
 				"dedup": []string{"sdfsf-range"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.labelValues,
@@ -541,7 +538,7 @@ func TestEndpoints(t *testing.T) {
 			params: map[string]string{
 				"name": "not!!!allowed",
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.series,
@@ -658,7 +655,7 @@ func TestEndpoints(t *testing.T) {
 		// Missing match[] query params in series requests.
 		{
 			endpoint: api.series,
-			errType:  errorBadData,
+			errType:  baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.series,
@@ -666,7 +663,7 @@ func TestEndpoints(t *testing.T) {
 				"match[]": []string{`test_metric2`},
 				"dedup":   []string{"sdfsf-series"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 		},
 		{
 			endpoint: api.series,
@@ -783,7 +780,7 @@ func TestEndpoints(t *testing.T) {
 		// Missing match[] query params in series requests.
 		{
 			endpoint: api.series,
-			errType:  errorBadData,
+			errType:  baseAPI.ErrorBadData,
 			method:   http.MethodPost,
 		},
 		{
@@ -792,7 +789,7 @@ func TestEndpoints(t *testing.T) {
 				"match[]": []string{`test_metric2`},
 				"dedup":   []string{"sdfsf-series"},
 			},
-			errType: errorBadData,
+			errType: baseAPI.ErrorBadData,
 			method:  http.MethodPost,
 		},
 	}
@@ -827,7 +824,7 @@ func TestEndpoints(t *testing.T) {
 
 			resp, _, apiErr := test.endpoint(req.WithContext(ctx))
 			if apiErr != nil {
-				if test.errType == errorNone {
+				if test.errType == baseAPI.ErrorNone {
 					t.Fatalf("Unexpected error: %s", apiErr)
 				}
 				if test.errType != apiErr.Typ {
@@ -835,7 +832,7 @@ func TestEndpoints(t *testing.T) {
 				}
 				return
 			}
-			if test.errType != errorNone {
+			if test.errType != baseAPI.ErrorNone {
 				t.Fatalf("Expected error of type %q but got none", test.errType)
 			}
 
@@ -846,82 +843,6 @@ func TestEndpoints(t *testing.T) {
 			return
 		}
 
-	}
-}
-
-func TestRespondSuccess(t *testing.T) {
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		Respond(w, "test", nil)
-	}))
-	defer s.Close()
-
-	resp, err := http.Get(s.URL)
-	if err != nil {
-		t.Fatalf("Error on test request: %s", err)
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	defer func() { testutil.Ok(t, resp.Body.Close()) }()
-	if err != nil {
-		t.Fatalf("Error reading response body: %s", err)
-	}
-
-	if resp.StatusCode != 200 {
-		t.Fatalf("Return code %d expected in success response but got %d", 200, resp.StatusCode)
-	}
-	if h := resp.Header.Get("Content-Type"); h != "application/json" {
-		t.Fatalf("Expected Content-Type %q but got %q", "application/json", h)
-	}
-
-	var res response
-	if err = json.Unmarshal([]byte(body), &res); err != nil {
-		t.Fatalf("Error unmarshaling JSON body: %s", err)
-	}
-
-	exp := &response{
-		Status: statusSuccess,
-		Data:   "test",
-	}
-	if !reflect.DeepEqual(&res, exp) {
-		t.Fatalf("Expected response \n%v\n but got \n%v\n", res, exp)
-	}
-}
-
-func TestRespondError(t *testing.T) {
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		RespondError(w, &ApiError{errorTimeout, errors.New("message")}, "test")
-	}))
-	defer s.Close()
-
-	resp, err := http.Get(s.URL)
-	if err != nil {
-		t.Fatalf("Error on test request: %s", err)
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	defer func() { testutil.Ok(t, resp.Body.Close()) }()
-	if err != nil {
-		t.Fatalf("Error reading response body: %s", err)
-	}
-
-	if want, have := http.StatusServiceUnavailable, resp.StatusCode; want != have {
-		t.Fatalf("Return code %d expected in error response but got %d", want, have)
-	}
-	if h := resp.Header.Get("Content-Type"); h != "application/json" {
-		t.Fatalf("Expected Content-Type %q but got %q", "application/json", h)
-	}
-
-	var res response
-	if err = json.Unmarshal([]byte(body), &res); err != nil {
-		t.Fatalf("Error unmarshaling JSON body: %s", err)
-	}
-
-	exp := &response{
-		Status:    statusError,
-		Data:      "test",
-		ErrorType: errorTimeout,
-		Error:     "message",
-	}
-	if !reflect.DeepEqual(&res, exp) {
-		t.Fatalf("Expected response \n%v\n but got \n%v\n", res, exp)
 	}
 }
 
@@ -1030,37 +951,6 @@ func TestParseDuration(t *testing.T) {
 	}
 }
 
-func TestOptionsMethod(t *testing.T) {
-	r := route.New()
-	api := &API{
-		gate: gate.NewKeeper(nil).NewGate(4),
-	}
-	api.Register(r, &opentracing.NoopTracer{}, log.NewNopLogger(), extpromhttp.NewNopInstrumentationMiddleware())
-
-	s := httptest.NewServer(r)
-	defer s.Close()
-
-	req, err := http.NewRequest("OPTIONS", s.URL+"/any_path", nil)
-	if err != nil {
-		t.Fatalf("Error creating OPTIONS request: %s", err)
-	}
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		t.Fatalf("Error executing OPTIONS request: %s", err)
-	}
-
-	if resp.StatusCode != http.StatusNoContent {
-		t.Fatalf("Expected status %d, got %d", http.StatusNoContent, resp.StatusCode)
-	}
-
-	for h, v := range corsHeaders {
-		if resp.Header.Get(h) != v {
-			t.Fatalf("Expected %q for header %q, got %q", v, h, resp.Header.Get(h))
-		}
-	}
-}
-
 func BenchmarkQueryResultEncoding(b *testing.B) {
 	var mat promql.Matrix
 	for i := 0; i < 1000; i++ {
@@ -1148,7 +1038,7 @@ func TestParseDownsamplingParamMillis(t *testing.T) {
 	}
 
 	for i, test := range tests {
-		api := API{
+		api := QueryAPI{
 			enableAutodownsampling: test.enableAutodownsampling,
 			gate:                   gate.NewKeeper(nil).NewGate(4),
 		}
